# API/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow API Docs",
    "description": "Documentacion interactiva completa de la API de CoffeeFlow: Autenticacion, Usuarios, Mesas, Productos, Pedidos, Cocina (ingredientes/recetas/inventario/alertas), Caja (turnos/tickets/pagos/cortes/gastos/compras) y Notificaciones.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Servidor Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/mesas": {
      "get": {
        "tags": [
          "Mesas"
        ],
        "summary": "Listar mesas",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "disponible",
                "ocupada",
                "reservada"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de mesas"
          }
        }
      },
      "post": {
        "tags": [
          "Mesas"
        ],
        "summary": "Crear mesa",
        "description": "Solo administrador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numero_mesa": {
                    "type": "integer",
                    "example": 9
                  },
                  "capacidad": {
                    "type": "integer",
                    "example": 6
                  },
                  "ubicacion": {
                    "type": "string",
                    "example": "Terraza"
                  },
                  "estado": {
                    "type": "string",
                    "enum": [
                      "disponible",
                      "ocupada",
                      "reservada"
                    ],
                    "example": "disponible"
                  }
                },
                "required": [
                  "numero_mesa"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mesa creada"
          }
        }
      }
    },
    "/api/mesas/{id_mesa}": {
      "get": {
        "tags": [
          "Mesas"
        ],
        "summary": "Obtener mesa",
        "parameters": [
          {
            "name": "id_mesa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mesa"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Mesas"
        ],
        "summary": "Actualizar mesa",
        "description": "Solo administrador.",
        "parameters": [
          {
            "name": "id_mesa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "capacidad": {
                    "type": "integer",
                    "example": 8
                  },
                  "ubicacion": {
                    "type": "string",
                    "example": "Terraza ampliada"
                  },
                  "numero_mesa": {
                    "type": "integer"
                  },
                  "estado": {
                    "type": "string",
                    "enum": [
                      "disponible",
                      "ocupada",
                      "reservada"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mesa actualizada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Mesas"
        ],
        "summary": "Eliminar mesa",
        "description": "Solo administrador.",
        "parameters": [
          {
            "name": "id_mesa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      }
    },
    "/api/mesas/{id_mesa}/estado": {
      "patch": {
        "tags": [
          "Mesas"
        ],
        "summary": "Cambiar estado de mesa",
        "parameters": [
          {
            "name": "id_mesa",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": {
                    "type": "string",
                    "enum": [
                      "disponible",
                      "ocupada",
                      "reservada"
                    ],
                    "example": "ocupada"
                  }
                },
                "required": [
                  "estado"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "400": {
            "description": "Estado invalido"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      }
    },
...
  }
}